+++ sw/airborne/firmwares/rotorcraft/guidance/guidance_hybrid.h
@@ -34,7 +34,7 @@
 
 extern int32_t guidance_hybrid_norm_ref_airspeed;
 extern float alt_pitch_gain;
-extern int32_t max_airspeed;
+extern float max_airspeed;
 extern int32_t wind_gain;
 extern int32_t horizontal_speed_gain;
 extern float max_turn_bank;
